code review tweaks
