+++ FileProcessor.Tests/BootstrapperTests.cs
@@ -30,34 +30,16 @@ namespace FileProcessor.Tests
 
             this.AddTestRegistrations(services, hostingEnvironment.Object);
             s.ConfigureContainer(services);
-
-            var g = Startup.Container.GetInstance<ISecurityServiceClient>();
-
-
+            
             Startup.Container.AssertConfigurationIsValid(AssertMode.Full);
         }
 
         private IConfigurationRoot SetupMemoryConfiguration()
         {
-            //Dictionary<String, String> configuration = new Dictionary<String, String>();
-
             IConfigurationBuilder builder = new ConfigurationBuilder();
 
-            //configuration.Add("ConnectionStrings:HealthCheck", "HeathCheckConnString");
-            //configuration.Add("SecurityConfiguration:Authority", "https://127.0.0.1");
-            //configuration.Add("EventStoreSettings:ConnectionString", "https://127.0.0.1:2113");
-            //configuration.Add("EventStoreSettings:ConnectionName", "UnitTestConnection");
-            //configuration.Add("EventStoreSettings:UserName", "admin");
-            //configuration.Add("EventStoreSettings:Password", "changeit");
-            //configuration.Add("AppSettings:UseConnectionStringConfig", "false");
-            //configuration.Add("AppSettings:SecurityService", "http://127.0.0.1");
-            //configuration.Add("AppSettings:MessagingServiceApi", "http://127.0.0.1");
-            //configuration.Add("AppSettings:EstateManagementApi", "http://127.0.0.1");
-            //configuration.Add("AppSettings:TransactionProcessorApi", "http://127.0.0.1");
-            //configuration.Add("AppSettings:DatabaseEngine", "SqlServer");
-
-            //builder.AddInMemoryCollection(configuration);
             builder.AddInMemoryCollection(TestData.DefaultAppSettings);
+
             return builder.Build();
         }
